fix: various other changes
A variety of changes were implemented in order to improve deployment. This includes updating kustomization files to add or remove certain Helm repositories and system-upgrade controllers. Network dependencies and settings in different Helm releases are updated to match the required configuration. A new file, corends.yaml, is added to the helm repositories and there are various changes in labels, configurations, and settings in various Kubernetes application files. Additionally, addresses and other configurations are updated or removed from the cluster-settings. The commit addresses various corrections or improvements in the Kubernetes setup and reflects the need for these changes for the proper functioning of the deployed applications and revisions.