feat: add Graphite Exporter for monitoring TrueNAS
This commit adds several new YAML configuration files required for deploying Graphite Exporter in a Kubernetes-based system. Graphite Exporter is used to monitor metrics on TrueNAS systems.

The yaml files specify how the Graphite exporter is to be deployed, its necessary configurations, and how it interfaces with the Kubernetes monitoring system. This includes the exporter's app config mappings, the helm release settings, the kustomizing configurations and dependencies for the FluxCD.

The changes are needed to improve the monitoring and alerting features, providing system administrators with more comprehensive and detailed information about the TrueNAS system's performance.