fix: jellyfin groups
This commit introduces updates to the Jellyfin helmrelease.yaml. Additional user groups for supplementalGroups field were added in the securityContext section. Now, Jellyfin runs using user group 44 and 1000, which resolves the permission issues experienced when trying to access certain system resources. This change was necessary to ensure the correct permissions are set for Jellyfin to function properly.