feat: enable health checks and dependencies for nvidia-gpu-operator
This commit changes the configuration of the nvidia-gpu-operator in the kubernetes setup. We enabled the waiting option, added a health check and set the nvidia-gpu-operator as a new dependency for the media/jellyfin app. The main motivation behind this commit is to ensure the stability of the app which heavily relies on GPUs and to prevent race conditions during deployment. Further, the namespace for the nvidia-gpu-operator is also changed to 'kube-system' for better isolation and management.