feat: update kube-prometheus-stack and velero configuration
This commit modifies the kube-prometheus-stack and velero configuration in the Kubernetes setup. In the kube-prometheus-stack configuration, unnecessary lines have been removed and additional scrape configurations have been added. These configurations help Prometheus scrape metrics from the various labeled nodes and subsequently relabel and reorganize them for better monitoring.

In the velero configuration, additional functionalities like nodeAgentPodMonitor and prometheusRule have been enabled along with the specification of alerts. The alerts 'VeleroBackupPartialFailures' and 'VeleroBackupFailures' are set up to actively monitor the backups and notify in case of failures.